Update by Apr.19th, 2026
Trivial update.
Update all dependencies.
Let QEMU to log exceptions in run_qemu.py
Rename EmulatorOps trait methods to mitigate name conflicts.
Remove likely_unlikely nightly feature. We will use cold_path.
Move advance_rip/inject_event into Vcpu methods.
Move vmread/vmwrite back into inline assembly to allow inline optimization.
Remove volatile XMM and replace with XSAVE state.
Implement more methods in VmcbOps trait.
Decouple EFER definition from SVM-Core.

+++ src/vt_core/hvcall.rs
@@ -35,7 +35,7 @@ impl VtVcpu
 		let gpr_state=&mut self.get_stack_top_mut().gpr_state;
 		let saved_state:GprState=GprState
 		{
-			rax:nrip as u64,
+			rax:nrip,
 			rcx:gflags as u64,
 			rdx:gpr_state.rsp,
 			rbx:gpr_state.rbx,
